refactor: extract sandbox spec tests
